backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
databases==0.6.1
alembic==1.11.1
pydantic==1.10.7
bcrypt==4.0.1
python-jose==3.3.0
scikit-learn==1.2.2
pandas==2.0.1
websockets==11.0.3
pytest==7.4.0
pytest-asyncio==0.21.0
python-multipart==0.0.6
python-dotenv==1.0.0
